+++ Assets/Global/Scripts/Announcements.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 882fa0bc0aebfad46b0cb71fbe4852b7
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
