+++ Assets/Global/Scripts/Announcements/DotIndicator.cs
@@ -1,0 +1,14 @@
+using UnityEngine;
+using UnityEngine.UI;
+
+public class DotIndicator : MonoBehaviour
+{
+    public Image backgroundImage;
+    public Image fillImage;
+
+    public DotIndicator(Image background, Image fill)
+    {
+        backgroundImage = background;
+        fillImage = fill;
+    }
+}
