+++ Assets/Global/Scripts/Announcements/DotIndicator.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: ce0e00b0fce7c324ea18bfb2b75a6145
